third_party/blink/web_tests/printing/page-break-orphans-and-widows-expected.txt

Test for orphans and widows

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1-1" is 1
PASS: page number of "page1-2" is 1
PASS: page number of "page1-3" is 1
PASS: page number of "page2" is 2
PASS: page number of "page3" is 3
PASS: page number of "page4" is 4
FAIL: expected page number of "page5" is 5. Was 4
PASS: page number of "page6-1" is 6
PASS: page number of "page6-2" is 6
PASS: page number of "page7" is 7
PASS: page number of "page8" is 8
FAIL: expected page number of "page9" is 9. Was 8

PASS successfullyParsed is true

TEST COMPLETE
